+++ src/HeapSort.java
@@ -1,5 +1,5 @@
 public class HeapSort implements SortingAlgorithm {
-    int basicOperations = 0;
+    long basicOperations = 0;
     int arrayLength = 0;
 
     @Override
@@ -41,7 +41,7 @@ public class HeapSort implements SortingAlgorithm {
     }
 
     @Override
-    public int getBasicOperations(){
+    public long getBasicOperations(){
         return basicOperations;
     }
 }
