+++ src/SortingAlgorithm.java
@@ -1,7 +1,7 @@
 public interface SortingAlgorithm {
-    int basicOperations = 0;
+    long basicOperations = 0;
     int arrayLength = 0;
     int[] sort(int[] array);
     void setArrayLength(int arrayLength);
-    int getBasicOperations();
+    long getBasicOperations();
 }
